Added calculateRadius method that is called right before requesting AQ data. Seems to work fine - gets data inside of a circle, not a rectangle, but Marrinan was fine with that.
Changed AQrequest to use JQuery, looks cleaner. Added a few checkboxes in the html file just for testing purposes. AQrequest can now send for multiple parameters based on theses textboxes, but I need to figure out how to simulate the "latest" data while using the "measurements" data. Have to play with date_from and date_to on the requestObject for that one.